refactor: event layout hook에서 share 책임 분리
